move graphing into own file
